+++ src/Sulu/Bundle/ContentBundle/Resources/public/js/components/content/preview/main.js
@@ -127,8 +127,8 @@ define(['app-config', 'config', 'websocket-manager'], function(AppConfig, Config
             },
 
             bindDomEvents = function() {
-                var changeFilter = 'input[type="checkbox"].preview-update, input[type="radio"].preview-update, select.preview-update',
-                    keyupFilter = '.preview-update:not(' + changeFilter + ')';
+                var changeFilter = '.preview-change-update, input[type="checkbox"].preview-update, input[type="radio"].preview-update, select.preview-update',
+                    keyupFilter = '.preview-update:not(' + changeFilter + ', .no-preview-update)';
 
                 this.sandbox.dom.on(this.formId, 'keyup', _.debounce(updateEvent.bind(this), this.config.delay), keyupFilter);
                 this.sandbox.dom.on(this.formId, 'change', updateEvent.bind(this), changeFilter);